+++ src/components/features/Theme/ThemeGoodsSection/index.tsx
@@ -1,52 +1,23 @@
 import styled from '@emotion/styled';
-import { useEffect, useState } from 'react';
+import { useMemo } from 'react';
 
-import { axiosInstance } from '@/api';
 import { DefaultGoodsItems } from '@/components/common/GoodsItem/Default';
 import { Container } from '@/components/common/layouts/Container';
 import { Grid } from '@/components/common/layouts/Grid';
 import { breakpoints } from '@/styles/variants';
 import type { GoodsData } from '@/types';
+import useFetch from '@/utils/api';
 
 type Props = {
   themeKey: string;
 };
 
-interface FetchState<T> {
-  isLoading: boolean;
-  isError: boolean;
-  data: T | null;
-}
-
 export const ThemeGoodsSection = ({ themeKey }: Props) => {
-  // const [goodsList, setGoodsList] = useState<GoodsData[]>([]);
-  const [fetchState, setFetchState] = useState<FetchState<GoodsData[]>>({
-    isLoading: true,
-    isError: false,
-    data: null,
-  });
-
-  useEffect(() => {
-    const fetchThemeGoods = async () => {
-      setFetchState({ isLoading: true, isError: false, data: [] });
-      try {
-        const response = await axiosInstance.get(`/api/v1/themes/${themeKey}/products`, {
-          params: {
-            maxResults: 20,
-          },
-        });
-        const { products } = response.data;
-        setFetchState({ isLoading: false, isError: false, data: products });
-      } catch (error) {
-        console.error('Error fetching theme goods:', error);
-        setFetchState({ isLoading: false, isError: true, data: null });
-      }
-    };
-
-    fetchThemeGoods();
-  }, [themeKey]);
-
-  const { isLoading, isError, data } = fetchState;
+  const params = useMemo(() => ({ maxResults: 20 }), []);
+  const { isLoading, isError, data } = useFetch<{ products: GoodsData[] }>(
+    `/api/v1/themes/${themeKey}/products`,
+    params,
+  );
 
   return (
     <Wrapper>
@@ -54,7 +25,7 @@ export const ThemeGoodsSection = ({ themeKey }: Props) => {
         <Description>로딩 중</Description>
       ) : isError ? (
         <Description>에러가 발생했습니다.</Description>
-      ) : data && data.length > 0 ? (
+      ) : data && data.products.length > 0 ? (
         <Container>
           <Grid
             columns={{
@@ -63,7 +34,7 @@ export const ThemeGoodsSection = ({ themeKey }: Props) => {
             }}
             gap={16}
           >
-            {data.map(({ id, imageURL, name, price, brandInfo }) => (
+            {data.products.map(({ id, imageURL, name, price, brandInfo }) => (
               <DefaultGoodsItems
                 key={id}
                 imageSrc={imageURL}
